Debug: Phenaki num_tokens was not consistent with codebook_size
