+++ phenaki_pytorch/__init__.py
@@ -1,4 +1,5 @@
-from phenaki_pytorch.phenaki_pytorch import Phenaki, CViViT, MaskGit, TokenCritic, make_video
+from .phenaki_pytorch import Phenaki, CViViT, TokenCritic, make_video
 
-from phenaki_pytorch.cvivit_trainer import CViViTTrainer
-from phenaki_pytorch.phenaki_trainer import PhenakiTrainer
+from .cvivit_trainer import CViViTTrainer
+from .mask_git import MaskGit
+from .phenaki_trainer import PhenakiTrainer
