+++ phenaki_pytorch/cvivit.py
@@ -258,13 +258,15 @@ class CViViT(nn.Module):
 
         super().__init__()
 
+        self.dim = dim
+        self.codebook_size = codebook_size
         self.image_size = pair(image_size)
         self.patch_size = pair(patch_size)
         patch_height, patch_width = self.patch_size
 
         self.temporal_patch_size = temporal_patch_size
 
-        self.spatial_rel_pos_bias = ContinuousPositionBias(dim = dim, heads = heads)
+        self.spatial_rel_pos_bias = ContinuousPositionBias(dim = self.dim, heads = heads)
 
         image_height, image_width = self.image_size
         assert (image_height % patch_height) == 0 and (image_width % patch_width) == 0
