+++ phenaki_pytorch/cvivit_trainer.py
@@ -222,11 +222,9 @@ class CViViTTrainer(nn.Module):
         self.vae.train()
 
         # logs
-
         logs = {}
 
         # update vae (generator)
-
         for _ in range(self.grad_accum_every):
             img = next(self.dl_iter)
             img = img.to(device)
@@ -250,7 +248,6 @@ class CViViTTrainer(nn.Module):
         # update discriminator
 
         self.accelerator.wait_for_everyone()
-
         if exists(self.vae.discr):
             self.discr_optim.zero_grad()
 
@@ -259,9 +256,7 @@ class CViViTTrainer(nn.Module):
                 img = img.to(device)
 
                 loss = self.vae(img, return_discr_loss = True)
-
                 self.accelerator.backward(loss / self.grad_accum_every)
-
                 accum_log(logs, {'discr_loss': loss.item() / self.grad_accum_every})
 
             if exists(self.discr_max_grad_norm):
@@ -274,14 +269,11 @@ class CViViTTrainer(nn.Module):
             self.print(f"{steps}: vae loss: {logs['loss']} - discr loss: {logs['discr_loss']}")
 
         # update exponential moving averaged generator
-
         self.accelerator.wait_for_everyone()
-
         if self.is_main and self.use_ema:
             self.ema_vae.update()
 
         # sample results every so often
-
         self.accelerator.wait_for_everyone()
 
         if self.is_main and not (steps % self.save_results_every):
@@ -295,16 +287,12 @@ class CViViTTrainer(nn.Module):
                 model.eval()
 
                 valid_data = next(self.valid_dl_iter)
-
                 is_video = valid_data.ndim == 5
-
                 valid_data = valid_data.to(device)
-
                 recons = model(valid_data, return_recons_only = True)
 
                 # if is video, save gifs to folder
                 # else save a grid of images
-
                 if is_video:
                     sampled_videos_path = self.results_folder / f'samples.{filename}'
                     (sampled_videos_path).mkdir(parents = True, exist_ok = True)
@@ -317,9 +305,7 @@ class CViViTTrainer(nn.Module):
 
                     imgs_and_recons = imgs_and_recons.detach().cpu().float().clamp(0., 1.)
                     grid = make_grid(imgs_and_recons, nrow = 2, normalize = True, value_range = (0, 1))
-
                     logs['reconstructions'] = grid
-
                     save_image(grid, str(self.results_folder / f'{filename}.png'))
 
             self.print(f'{steps}: saving to {str(self.results_folder)}')
@@ -327,7 +313,6 @@ class CViViTTrainer(nn.Module):
         # save model every so often
 
         self.accelerator.wait_for_everyone()
-
         if self.is_main and not (steps % self.save_model_every):
             state_dict = self.vae.state_dict()
             model_path = str(self.results_folder / f'vae.{steps}.pt')
